update version to 1.14.0

+++ PointSDK.iOS/Structs.cs
@@ -15,7 +15,7 @@ namespace PointSDK.iOS
     }
 
     [Native]
-    public enum BDAuthorizationLevel : nint
+    public enum BDAuthorizationLevel : long
     {
         Always,
         WhenInUse
